+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Using exponential backoff for retrying commit
 - Added complete ViurException hierarchy using the status error field and a mapping to our destinct errors.
 - silencing out logging output from requests and urllib3
+- introducing a config var 'verbose_error_codes' for selecting which error status should trigger a verbose message output on stdout/stderr 
 
 ### Changed